Remove invalidGrantedUsedHealthConnectPermissionsAreListed test
HealthConnect is improving behavior to remove any granted permissions for apps that do not define the required VIEW_PERMISSION_USAGE intent filter. As a result, this test no longer reliably passes, since the permissions get revoked and the behavior that the test is trying to verify no longer works as before.

Let me know if we still think this is important to keep, and we can discuss how.

Change-Id: I42678dc81b87d5668fc8d62bc16e0149d341a886
Flag: NONE test only
Bug: 370449451
Relnote: N/A