Enable secondary_user_on_secondary_display for CtsPermissionUiTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Most of the change is to specify the display ID in the UI selector to
accommodate the passenger display. Another change is to use a
different way to scroll the view for passenger as the previous approach
is deprecated and doesn't support scrolling on the passenger display.

Fixes: 355564081
Test: atest CtsPermissionUiTestCases
Test: atest --user-type secondary_user CtsPermissionUiTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsPermissionUiTestCases
Relnote: N/A
Flag: TEST_ONLY
LOW_COVERAGE_REASON=TEST_ONLY

Change-Id: I583163a2be2c3476d02d2357a93a9e26d8ba3bb9