Update minSdkVersion=Baklava for health permissions background permission test
When we enabled the flag, verifyBackgroundPropertiesForHealthPermissions() CTS test on V image fails. We're able to reproduce this locally, and we found the reason is: the V-image does not have this flag(flag not enabled on device), but the test is built from head where the flag is enabled(flag enabled for test).

To fix this, we need to update the minSdkVerision so that it does not run on a device that has inconsistent flag values.

Test: https://android-build.corp.google.com/builds/abtd/run/L10800030007846189
Relnote: Granular health permission phase one
Bug: 364638912
Bug: 379377663
Flag: android.permission.flags.replace_body_sensor_permission_enabled
LOW_COVERAGE_REASON=TEST_ONLY
Change-Id: I74c431778d0069bd40f9c4857b4a72521a511c92