Implement preference highlighting for expressive design.
This change uses Drawable#setColorFilter to highlight expressive
preferences. Unfortunately this doesn't work to highlight non-expressive
preferences so the original functionality (calling
View#setBackgroundColor) is retained.

This change started as a fresh fork of the class in Settings it is based
off of, and has some upstream changes that focus the target preference
as well as highlighting it.

Bug: 378433878
Test: Manual
Flag: NONE Changes are protected by system-wide prop
Relnote: Update Safety Center preference highlighting
LOW_COVERAGE_REASON=b/385322463

Change-Id: If11acd1b45ec43466d7227fc9ca9d9be5d233fcf