Move to Kotlin requireNotNull() from Preconditions.checkNotNull()
We are in Kotlin so the Kotlin way can be preferred.

Somehow the JSpecify @NonNull from Preconditions.checkNotNull()
isn't being recognized as non-null by the Kotlin compiler used for
the code here and that's blocking AndroidX prebuilt drop, so let's
get out of the way for that first, while improving our code a tiny
bit.

Bug: 385912424
Flag: EXEMPT bugfix
Relnote: N/A
Test: build
Change-Id: I42504a24a8544b56dbbbf5e0cd707355304717de