EnhancedConfirmationManagerTest: Use non-deprecated permission group in test.
The SENSORS permission group has been deprecated and will be removed (so the UI will no longer prompt for it). Update this test to use a different unrestricted permission/group. Camera was chosen as another unrestricted permission but we can swap it for another if that would be preferred.

Bug: 388616482
Change-Id: Ib8f63ec0e77d7dfef9ed5ab49ba036c9df88251d
Test: atest CtsPermissionUiTestCases:android.permissionui.cts.EnhancedConfirmationManagerTest
Flag: TEST_ONLY
Relnote: Fix CTS test with BODY_SENSORS deprecation.