[ECM] Wait for MODE_DEFAULT (RoleManagerTest)
After each package installation, the ACCESS_RESTRICTED_SETTINGS op
should get set to MODE_DEFAULT. But, it happens asynchronously, so we
need to wait for it to occur.

And in this case, if we don't wait, then it may happen after we set it,
interfering with our test.

Bug: 384418063
Flag: android.permission.flags.enhanced_confirmation_mode_apis_enabled
Test: atest CtsPermissionUiTestCases:android.permissionui.cts.EnhancedConfirmationManagerTest
LOW_COVERAGE_REASON=TEST_ONLY
Relnote: N/A
Change-Id: Ida525932b195c358e68854cc2d25641766f516c3