[ECM] Wait for MODE_DEFAULT (AppPermissionTest)
After each package installation, the ACCESS_RESTRICTED_SETTINGS op
should get set to MODE_DEFAULT. But, it happens asynchronously, so we
need to wait for it to occur.

And in this case, if we don't wait, then the test may see MODE_ALLOWED
as the app op mode, which will break the test.

Bug: 384418063
Flag: android.permission.flags.enhanced_confirmation_mode_apis_enabled
Test: atest CtsPermissionUiTestCases:android.permissionui.cts.AppPermissionTest
LOW_COVERAGE_REASON=TEST_ONLY
Relnote: N/A
Change-Id: Iee720707f72265d8739fc250026831c504c98270