[ECM] Add sleep and mark as FlakyTest
For these tests, on certain devices (AOSP+cuttlefish), the grant dialog
appears to hang even after UiAutomator clicks on a button. The root
cause for this is unknown, but logs seem to tell us that the clicks are
are accidentally landing on RequestPermissionsActivity (instead of
GrantPermissionsActivity), where the button doesn't exist.

This problem happens *even if we tell UiAutomator to first wait for the
target button to visible on-screen and clickable*. Likely this issue has
something to do with the fact that we're showing two popups: one for
ECM and one for grants, and both are part of RequestPermissionsActivity;
somehow UiAutomator thinks the grant dialog button is visible in both
cases.

To mitigate this, add a short sleep which seems to help. To prevent this
from possibly still blocking presubmit, add @FlakyTest which excludes
the test from presubmit.

Bug: 387927331
Bug: 390440965
Test: manual
Flag: TEST_ONLY
Relnote: N/A
LOW_COVERAGE_REASON=TEST_ONLY
Change-Id: I68b72bcebc4ebd2e17c72f9232801278b16e607b