Give clickable icon a contentDescription
Give this icon a contentDescription to satisfy accessibility
requirements.

Depending on the situation, this icon might be an Info icon or a
Settings icon, so we'll need to set the contentDescription dynamically.

Since AppPermissionFragment has been split into
LegacyAppPermissionFragment and AppPermissionFragment, we need to fix
this separately in both fragments.

Fix: 379788955
Test: manual (enable talkback, click icon, inspect with HSV)
Relnote: N/A
Flag: EXEMPT bugfix
Change-Id: Id4348c5a24d2652e9c9da9db37f2d8f8e27f0fcb