[E2E] Fix edge-to-edge
Fix edge-to-edge in PermissionController.

We recently enabled edge-to-edge (go/e2e-2024-1p) in the
Permission module, and partially resolved some margin issues by
adding `fitsSystemWindows="true"` to the permissions_frame layout.

But even after this fix, there still appeared to be an issue with the
collapsing toolbar. (It was as if `fitsSystemWindows="true"` had fixed
everything except for the collapsing toolbar.)

This collapsing toolbar issue is now fixed here by moving the
`fitsSystemWindows=true` up to the nav_host_fragment layout instead of
the permissions_frame layout.

Here's a rough portayal of how these layouts are nested:

	R.layout.nav_host_fragment {
	    R.layout.collapsing_toolbar_base_layout {
		<AppBarLayout>
		    <CollapsingToolbarLayout .../>
		</AppBarLayout>
		R.layout.permissions_frame { ... }
	    }
	}

Bug: 380055900
Test: atest CtsPermissionUiTestCases
Relnote: N/A
Flag: EXEMPT bugfix
Change-Id: I4fc71837d44e42342d6865e2a146f3848375ebe1