Refactor default app implementation
So that it can better accommodate new features.

Default holders logic is extracted into SignedPackage(Utils) for reuse,
a RoleApplicationItem is added to replace the previous Pair, the
LiveData transformation functions are made generic, and
DefaultAppChildFragment preference code is also refactored to be more
reusable.

Bug: 388234667
Relnote: N/A
Flag: EXEMPT bugfix
Test: presubmit
Change-Id: I3285e868c6d7457dfd42bbe12c741e5cae2f78f5