Add recommended category for default app page
For assistant, it can be enabled by overlaying PermissionController
config_recommendedAssistants with a list of package names, optionally
with signing certificate digests if the app needs to be a non-system
app, similar to the config for default holders.

This is not available for the request role dialog yet and thus still
can't be used with a role that allows requesting.

Bug: 388234667
Relnote: N/A
Flag: com.android.permission.flags.default_apps_recommendation_enabled
Test: manual
Change-Id: I10a387c31362a126ee6264915f40f14926e06910