RESTRICT AUTOMERGE Fix: call `shouldSkipPermission` for platform namespace permissions
Change I446318e449547f0393124da23655fbe38ae52485 introduced a fix to
bypass the check of the DYNAMIC_INSTRUMENTATION permission on Android 15
devices.

Since this permission is sometimes added via mainline update, it is not
necessarily in the `expectedPermissions` array. This happens if the
permission didn't exist in the OS prior to the mainline update.

RESTRICT AUTOMERGE
Bug: 406917964
Test: CtsPermissionPolicyTestCases
Flag: EXEMPT test change only
Change-Id: I0f2ba348c8e44aa4049f731ffd3ee1a86b7fb432