Skip test that uses DeviceConfig.
Following on from Ic7b68809c9e2001582d10bb183e351d6572a2bbb there is no
reliable way to detect if the device actually is or isn't using
DeviceConfig flag to control Safety Center. We assume always that it
isn't, but in this test that assumption (if it is not true) causes an
error where we turn Safety Center off and then leave it stuck off.

Test: atest CtsSafetyCenterTestCases
Bug: 423822838
Change-Id: Id6da9f2d0ce934a4e1bde349d8cbb4e4ad151d92